ADD GIMP, PROXY AND WORDPRESS TOOLS & SITE WORDPRESS REPO AS SUBMODULES
We're going to integrate the tools-gimp, tools-proxy and tools-wordpress
repos as submodules with this repo's Docker Compose project. Then we can
use the tools that they provide. The newopera-wordpress repo is also
added as a submodule of this repo, as this is the website that we want
to develop and test on the client, using Docker Compose.