ADD VARILINK WORDPRESS PLUGINS LIBRARY AS A SUBMODULE
I've done this to get access to the varilink-log-writer plugin for
debugging purposes. It occurs to me that this is a different use case to
the one in which the WordPress site itself is using one of the plugins
in this library. That use case would lead me to include the library as a
submodule of the WordPress sites repository whereas this use case leads
me to include it as a submodule of the associated Docker Compose
repository.

I shall have to think in future about whether to partition that library
into one for dev tools (only varilink-log-writer falls into that
category right now) and one for plugins used in the WordPress site. Or
perhaps move to per plugin repositories? TBC.